+++ game.h
@@ -1,6 +1,6 @@
-void new_game_board(short (*board)[4]);
-void print_board(short (*board)[4]);
-int is_equal_board(short (*board1)[4], short (*board2)[4]);
-void add_random_tile(short tile, short (*board)[4]);
-void add_random_tile_start(short (*board)[4]);
-int frequence(short (*board)[4], short number);
+void new_game_board(int (*board)[4]);
+void print_board(int (*board)[4]);
+int is_equal_board(int (*board1)[4], int (*board2)[4]);
+void add_random_tile(int tile, int (*board)[4]);
+void add_random_tile_start(int (*board)[4]);
+int frequence(int (*board)[4], int number);
